d2fmt: lowercase reserved keywords
