+++ d2graph/d2graph.go
@@ -268,8 +268,8 @@ func (s *Style) Apply(key, value string) error {
 		if s.FillPattern == nil {
 			break
 		}
-		if !go2.Contains(FillPatterns, strings.ToLower(value)) {
-			return fmt.Errorf(`expected "fill-pattern" to be one of: %s`, strings.Join(FillPatterns, ", "))
+		if !go2.Contains(d2ast.FillPatterns, strings.ToLower(value)) {
+			return fmt.Errorf(`expected "fill-pattern" to be one of: %s`, strings.Join(d2ast.FillPatterns, ", "))
 		}
 		s.FillPattern.Value = value
 	case "stroke-width":
@@ -409,8 +409,8 @@ func (s *Style) Apply(key, value string) error {
 		if s.TextTransform == nil {
 			break
 		}
-		if !go2.Contains(textTransforms, strings.ToLower(value)) {
-			return fmt.Errorf(`expected "text-transform" to be one of (%s)`, strings.Join(textTransforms, ", "))
+		if !go2.Contains(d2ast.TextTransforms, strings.ToLower(value)) {
+			return fmt.Errorf(`expected "text-transform" to be one of (%s)`, strings.Join(d2ast.TextTransforms, ", "))
 		}
 		s.TextTransform.Value = value
 	default:
@@ -667,7 +667,7 @@ func (obj *Object) HasChild(ids []string) (*Object, bool) {
 		return obj, true
 	}
 	if len(ids) == 1 && ids[0] != "style" {
-		_, ok := ReservedKeywords[ids[0]]
+		_, ok := d2ast.ReservedKeywords[ids[0]]
 		if ok {
 			return obj, true
 		}
@@ -818,9 +818,9 @@ func (obj *Object) EnsureChild(ida []string) *Object {
 		return obj
 	}
 
-	_, is := ReservedKeywordHolders[ida[0]]
+	_, is := d2ast.ReservedKeywordHolders[ida[0]]
 	if len(ida) == 1 && !is {
-		_, ok := ReservedKeywords[ida[0]]
+		_, ok := d2ast.ReservedKeywords[ida[0]]
 		if ok {
 			return obj
 		}
@@ -1120,7 +1120,7 @@ func (obj *Object) IsConstantNear() bool {
 	if isKey {
 		return false
 	}
-	_, isConst := NearConstants[keyPath[0]]
+	_, isConst := d2ast.NearConstants[keyPath[0]]
 	return isConst
 }
 
@@ -1235,7 +1235,7 @@ func (e *Edge) AbsID() string {
 func (obj *Object) Connect(srcID, dstID []string, srcArrow, dstArrow bool, label string) (*Edge, error) {
 	for _, id := range [][]string{srcID, dstID} {
 		for _, p := range id {
-			if _, ok := ReservedKeywords[p]; ok {
+			if _, ok := d2ast.ReservedKeywords[p]; ok {
 				return nil, errors.New("cannot connect to reserved keyword")
 			}
 		}
@@ -1428,12 +1428,12 @@ func (g *Graph) SetDimensions(mtexts []*d2target.MText, ruler *textmeasure.Ruler
 		// user-specified label/icon positions
 		if obj.HasLabel() && obj.Attributes.LabelPosition != nil {
 			scalar := *obj.Attributes.LabelPosition
-			position := LabelPositionsMapping[scalar.Value]
+			position := d2ast.LabelPositionsMapping[scalar.Value]
 			obj.LabelPosition = go2.Pointer(position.String())
 		}
 		if obj.Icon != nil && obj.Attributes.IconPosition != nil {
 			scalar := *obj.Attributes.IconPosition
-			position := LabelPositionsMapping[scalar.Value]
+			position := d2ast.LabelPositionsMapping[scalar.Value]
 			obj.IconPosition = go2.Pointer(position.String())
 		}
 
@@ -1671,205 +1671,6 @@ func (g *Graph) Texts() []*d2target.MText {
 
 func Key(k *d2ast.KeyPath) []string {
 	return d2format.KeyPath(k)
-}
-
-// All reserved keywords. See init below.
-var ReservedKeywords map[string]struct{}
-
-// Non Style/Holder keywords.
-var SimpleReservedKeywords = map[string]struct{}{
-	"label":          {},
-	"desc":           {},
-	"shape":          {},
-	"icon":           {},
-	"constraint":     {},
-	"tooltip":        {},
-	"link":           {},
-	"near":           {},
-	"width":          {},
-	"height":         {},
-	"direction":      {},
-	"top":            {},
-	"left":           {},
-	"grid-rows":      {},
-	"grid-columns":   {},
-	"grid-gap":       {},
-	"vertical-gap":   {},
-	"horizontal-gap": {},
-	"class":          {},
-	"vars":           {},
-}
-
-// ReservedKeywordHolders are reserved keywords that are meaningless on its own and must hold composites
-var ReservedKeywordHolders = map[string]struct{}{
-	"style":            {},
-	"source-arrowhead": {},
-	"target-arrowhead": {},
-}
-
-// CompositeReservedKeywords are reserved keywords that can hold composites
-var CompositeReservedKeywords = map[string]struct{}{
-	"classes":    {},
-	"constraint": {},
-	"label":      {},
-	"icon":       {},
-}
-
-// StyleKeywords are reserved keywords which cannot exist outside of the "style" keyword
-var StyleKeywords = map[string]struct{}{
-	"opacity":       {},
-	"stroke":        {},
-	"fill":          {},
-	"fill-pattern":  {},
-	"stroke-width":  {},
-	"stroke-dash":   {},
-	"border-radius": {},
-
-	// Only for text
-	"font":           {},
-	"font-size":      {},
-	"font-color":     {},
-	"bold":           {},
-	"italic":         {},
-	"underline":      {},
-	"text-transform": {},
-
-	// Only for shapes
-	"shadow":        {},
-	"multiple":      {},
-	"double-border": {},
-
-	// Only for squares
-	"3d": {},
-
-	// Only for edges
-	"animated": {},
-	"filled":   {},
-}
-
-// TODO maybe autofmt should allow other values, and transform them to conform
-// e.g. left-center becomes center-left
-var NearConstantsArray = []string{
-	"top-left",
-	"top-center",
-	"top-right",
-
-	"center-left",
-	"center-right",
-
-	"bottom-left",
-	"bottom-center",
-	"bottom-right",
-}
-var NearConstants map[string]struct{}
-
-// LabelPositionsArray are the values that labels and icons can set `near` to
-var LabelPositionsArray = []string{
-	"top-left",
-	"top-center",
-	"top-right",
-
-	"center-left",
-	"center-center",
-	"center-right",
-
-	"bottom-left",
-	"bottom-center",
-	"bottom-right",
-
-	"outside-top-left",
-	"outside-top-center",
-	"outside-top-right",
-
-	"outside-left-top",
-	"outside-left-center",
-	"outside-left-bottom",
-
-	"outside-right-top",
-	"outside-right-center",
-	"outside-right-bottom",
-
-	"outside-bottom-left",
-	"outside-bottom-center",
-	"outside-bottom-right",
-}
-var LabelPositions map[string]struct{}
-
-// convert to label.Position
-var LabelPositionsMapping = map[string]label.Position{
-	"top-left":   label.InsideTopLeft,
-	"top-center": label.InsideTopCenter,
-	"top-right":  label.InsideTopRight,
-
-	"center-left":   label.InsideMiddleLeft,
-	"center-center": label.InsideMiddleCenter,
-	"center-right":  label.InsideMiddleRight,
-
-	"bottom-left":   label.InsideBottomLeft,
-	"bottom-center": label.InsideBottomCenter,
-	"bottom-right":  label.InsideBottomRight,
-
-	"outside-top-left":   label.OutsideTopLeft,
-	"outside-top-center": label.OutsideTopCenter,
-	"outside-top-right":  label.OutsideTopRight,
-
-	"outside-left-top":    label.OutsideLeftTop,
-	"outside-left-center": label.OutsideLeftMiddle,
-	"outside-left-bottom": label.OutsideLeftBottom,
-
-	"outside-right-top":    label.OutsideRightTop,
-	"outside-right-center": label.OutsideRightMiddle,
-	"outside-right-bottom": label.OutsideRightBottom,
-
-	"outside-bottom-left":   label.OutsideBottomLeft,
-	"outside-bottom-center": label.OutsideBottomCenter,
-	"outside-bottom-right":  label.OutsideBottomRight,
-}
-
-var FillPatterns = []string{
-	"none",
-	"dots",
-	"lines",
-	"grain",
-	"paper",
-}
-
-var textTransforms = []string{"none", "uppercase", "lowercase", "capitalize"}
-
-// BoardKeywords contains the keywords that create new boards.
-var BoardKeywords = map[string]struct{}{
-	"layers":    {},
-	"scenarios": {},
-	"steps":     {},
-}
-
-func init() {
-	ReservedKeywords = make(map[string]struct{})
-	for k, v := range SimpleReservedKeywords {
-		ReservedKeywords[k] = v
-	}
-	for k, v := range StyleKeywords {
-		ReservedKeywords[k] = v
-	}
-	for k, v := range ReservedKeywordHolders {
-		CompositeReservedKeywords[k] = v
-	}
-	for k, v := range BoardKeywords {
-		CompositeReservedKeywords[k] = v
-	}
-	for k, v := range CompositeReservedKeywords {
-		ReservedKeywords[k] = v
-	}
-
-	NearConstants = make(map[string]struct{}, len(NearConstantsArray))
-	for _, k := range NearConstantsArray {
-		NearConstants[k] = struct{}{}
-	}
-
-	LabelPositions = make(map[string]struct{}, len(LabelPositionsArray))
-	for _, k := range LabelPositionsArray {
-		LabelPositions[k] = struct{}{}
-	}
 }
 
 func (g *Graph) GetBoard(name string) *Graph {
