+++ d2oracle/edit.go
@@ -372,7 +372,7 @@ func _set(g *d2graph.Graph, baseAST *d2ast.Map, key string, tag, value *string) 
 	if mk.Key != nil {
 		found := true
 		for _, idel := range d2graph.Key(mk.Key) {
-			_, ok := d2graph.ReservedKeywords[idel]
+			_, ok := d2ast.ReservedKeywords[idel]
 			if ok {
 				reserved = true
 				break
@@ -544,7 +544,7 @@ func _set(g *d2graph.Graph, baseAST *d2ast.Map, key string, tag, value *string) 
 		attrs = edge.Attributes
 
 		if mk.EdgeKey != nil {
-			if _, ok := d2graph.ReservedKeywords[mk.EdgeKey.Path[0].Unbox().ScalarString()]; !ok {
+			if _, ok := d2ast.ReservedKeywords[mk.EdgeKey.Path[0].Unbox().ScalarString()]; !ok {
 				return errors.New("edge key must be reserved")
 			}
 			reserved = true
@@ -856,7 +856,7 @@ func appendMapKey(m *d2ast.Map, mk *d2ast.Key) {
 	if len(m.Nodes) == 1 &&
 		mk.Key != nil &&
 		len(mk.Key.Path) > 0 {
-		_, ok := d2graph.ReservedKeywords[mk.Key.Path[0].Unbox().ScalarString()]
+		_, ok := d2ast.ReservedKeywords[mk.Key.Path[0].Unbox().ScalarString()]
 		if ok {
 			// Allow one line reserved key (like shape) maps.
 			// TODO: This needs to be smarter as certain keys are only reserved in context.
@@ -1077,7 +1077,7 @@ func hoistRefChildren(g *d2graph.Graph, key *d2ast.KeyPath, ref d2graph.Referenc
 			continue
 		}
 		if n.MapKey.Key != nil {
-			_, ok := d2graph.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
+			_, ok := d2ast.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
 			if ok {
 				continue
 			}
@@ -1130,7 +1130,7 @@ func renameConflictsToParent(g *d2graph.Graph, key *d2ast.KeyPath) (*d2graph.Gra
 					continue
 				}
 				if n.MapKey.Key != nil {
-					_, ok := d2graph.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
+					_, ok := d2ast.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
 					if ok {
 						continue
 					}
@@ -1150,7 +1150,7 @@ func renameConflictsToParent(g *d2graph.Graph, key *d2ast.KeyPath) (*d2graph.Gra
 				absKey.Path = append(absKey.Path, k.Path[0])
 				absKeys = append(absKeys, absKey)
 			}
-		} else if _, ok := d2graph.ReservedKeywords[ref.Key.Path[len(ref.Key.Path)-1].Unbox().ScalarString()]; !ok {
+		} else if _, ok := d2ast.ReservedKeywords[ref.Key.Path[len(ref.Key.Path)-1].Unbox().ScalarString()]; !ok {
 			absKey, err := d2parser.ParseKey(ref.ScopeObj.AbsID())
 			if err != nil {
 				absKey = &d2ast.KeyPath{}
@@ -1254,7 +1254,7 @@ func deleteReserved(g *d2graph.Graph, boardPath []string, baseAST *d2ast.Map, mk
 		}
 		targetKey = mk.EdgeKey
 	}
-	_, ok := d2graph.ReservedKeywords[targetKey.Path[len(targetKey.Path)-1].Unbox().ScalarString()]
+	_, ok := d2ast.ReservedKeywords[targetKey.Path[len(targetKey.Path)-1].Unbox().ScalarString()]
 	if !ok {
 		return g, nil
 	}
@@ -1297,7 +1297,7 @@ func deleteReserved(g *d2graph.Graph, boardPath []string, baseAST *d2ast.Map, mk
 	imported := false
 	parts := d2graph.Key(targetKey)
 	for i, id := range parts {
-		_, ok := d2graph.ReservedKeywords[id]
+		_, ok := d2ast.ReservedKeywords[id]
 		if ok {
 			if id == "style" {
 				isNestedKey = true
@@ -1475,7 +1475,7 @@ func deleteObject(g *d2graph.Graph, baseAST *d2ast.Map, key *d2ast.KeyPath, obj 
 			}
 			ref.Key.Path = append(ref.Key.Path[:ref.KeyPathIndex], ref.Key.Path[ref.KeyPathIndex+1:]...)
 			withoutSpecial := go2.Filter(ref.Key.Path, func(x *d2ast.StringBox) bool {
-				_, isReserved := d2graph.ReservedKeywords[x.Unbox().ScalarString()]
+				_, isReserved := d2ast.ReservedKeywords[x.Unbox().ScalarString()]
 				isSpecial := isReserved || x.Unbox().ScalarString() == "_"
 				return !isSpecial
 			})
@@ -1494,7 +1494,7 @@ func deleteObject(g *d2graph.Graph, baseAST *d2ast.Map, key *d2ast.KeyPath, obj 
 				for i := 0; i < len(ref.MapKey.Value.Map.Nodes); i++ {
 					n := ref.MapKey.Value.Map.Nodes[i]
 					if n.MapKey != nil && n.MapKey.Key != nil {
-						_, ok := d2graph.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
+						_, ok := d2ast.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
 						if ok {
 							deleteFromMap(ref.MapKey.Value.Map, n.MapKey)
 							i--
@@ -1661,7 +1661,7 @@ func Rename(g *d2graph.Graph, boardPath []string, key, newName string) (_ *d2gra
 		mk2.Key = mk.Key
 		mk = mk2
 	} else {
-		_, ok := d2graph.ReservedKeywords[newName]
+		_, ok := d2ast.ReservedKeywords[newName]
 		if ok {
 			return nil, "", fmt.Errorf("cannot rename to reserved keyword: %#v", newName)
 		}
@@ -1686,7 +1686,7 @@ func Rename(g *d2graph.Graph, boardPath []string, key, newName string) (_ *d2gra
 
 func trimReservedSuffix(path []*d2ast.StringBox) []*d2ast.StringBox {
 	for i, p := range path {
-		if _, ok := d2graph.ReservedKeywords[p.Unbox().ScalarString()]; ok {
+		if _, ok := d2ast.ReservedKeywords[p.Unbox().ScalarString()]; ok {
 			return path[:i]
 		}
 	}
@@ -1933,9 +1933,9 @@ func move(g *d2graph.Graph, boardPath []string, key, newKey string, includeDesce
 			ida = resolvedIDA
 		}
 		// e.g. "a.b.shape: circle"
-		_, endsWithReserved := d2graph.ReservedKeywords[ida[len(ida)-1]]
+		_, endsWithReserved := d2ast.ReservedKeywords[ida[len(ida)-1]]
 		ida = go2.Filter(ida, func(x string) bool {
-			_, ok := d2graph.ReservedKeywords[x]
+			_, ok := d2ast.ReservedKeywords[x]
 			return !ok
 		})
 
@@ -1976,7 +1976,7 @@ func move(g *d2graph.Graph, boardPath []string, key, newKey string, includeDesce
 								continue
 							}
 							if n.MapKey.Key != nil {
-								_, ok := d2graph.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
+								_, ok := d2ast.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
 								if ok {
 									detachedMK.Value.Map.Nodes = append(detachedMK.Value.Map.Nodes, n)
 								}
@@ -2222,7 +2222,7 @@ func filterReserved(value d2ast.ValueBox) (with, without d2ast.ValueBox) {
 				forWithout = append(forWithout, n)
 				continue
 			}
-			_, ok := d2graph.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
+			_, ok := d2ast.ReservedKeywords[n.MapKey.Key.Path[0].Unbox().ScalarString()]
 			if !ok {
 				flushComments(&forWithout)
 				forWithout = append(forWithout, n)
@@ -2299,7 +2299,7 @@ func updateNear(prevG, g *d2graph.Graph, from, to *string, includeDescendants bo
 			}
 			if n.MapKey.Key.Path[len(n.MapKey.Key.Path)-1].Unbox().ScalarString() == "near" {
 				k := n.MapKey.Value.ScalarBox().Unbox().ScalarString()
-				if _, ok := d2graph.NearConstants[k]; ok {
+				if _, ok := d2ast.NearConstants[k]; ok {
 					continue
 				}
 				if strings.EqualFold(k, *from) && to == nil {
@@ -2925,7 +2925,7 @@ func DeleteIDDeltas(g *d2graph.Graph, boardPath []string, key string) (deltas ma
 	if mk.Key != nil {
 		ida := d2graph.Key(mk.Key)
 		// Deleting a reserved field cannot possibly have any deltas
-		if _, ok := d2graph.ReservedKeywords[ida[len(ida)-1]]; ok {
+		if _, ok := d2ast.ReservedKeywords[ida[len(ida)-1]]; ok {
 			return nil, nil
 		}
 
@@ -3256,7 +3256,7 @@ func getMostNestedRefs(obj *d2graph.Object) []d2graph.Reference {
 
 func filterReservedPath(path []*d2ast.StringBox) (filtered []*d2ast.StringBox) {
 	for _, box := range path {
-		if _, ok := d2graph.ReservedKeywords[strings.ToLower(box.Unbox().ScalarString())]; ok {
+		if _, ok := d2ast.ReservedKeywords[strings.ToLower(box.Unbox().ScalarString())]; ok {
 			return
 		}
 		filtered = append(filtered, box)
